chore(*) dataplane overiew filtering with pagination
