+++ pkg/core/resources/apis/mesh/dataplane_overview_test.go
@@ -2,10 +2,8 @@ package mesh_test
 
 import (
 	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
 
-	"github.com/kumahq/kuma/api/mesh/v1alpha1"
 	. "github.com/kumahq/kuma/pkg/core/resources/apis/mesh"
 	"github.com/kumahq/kuma/pkg/test/resources/model"
 )
@@ -44,112 +42,6 @@ var _ = Describe("DataplaneOverview", func() {
 			Expect(overviews.Items[0].Spec.DataplaneInsight).To(Equal(&insights.Items[0].Spec))
 			Expect(overviews.Items[1].Spec.Dataplane).To(Equal(&dataplanes.Items[1].Spec))
 			Expect(overviews.Items[1].Spec.DataplaneInsight).To(BeNil())
-		})
-	})
-
-	Describe("RetainMatchingTags", func() {
-		overviews := DataplaneOverviewResourceList{
-			Items: []*DataplaneOverviewResource{
-				{
-					Spec: v1alpha1.DataplaneOverview{
-						Dataplane: &v1alpha1.Dataplane{
-							Networking: &v1alpha1.Dataplane_Networking{
-								Inbound: []*v1alpha1.Dataplane_Networking_Inbound{
-									{
-										Tags: map[string]string{
-											"kuma.io/service": "mobile",
-											"version":         "v1",
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		}
-		type testCase struct {
-			tags     map[string]string
-			expected DataplaneOverviewResourceList
-		}
-		DescribeTable("should retain overviews", func(given testCase) {
-			// when
-			overviews.RetainMatchingTags(given.tags)
-
-			// then
-			Expect(overviews).To(Equal(given.expected))
-		},
-			Entry("should retain all with empty map", testCase{
-				tags:     map[string]string{},
-				expected: overviews,
-			}),
-			Entry("should retain with one matching tag", testCase{
-				tags:     map[string]string{"kuma.io/service": "mobile"},
-				expected: overviews,
-			}),
-			Entry("should retain with matching all tags", testCase{
-				tags:     map[string]string{"kuma.io/service": "mobile", "version": "v1"},
-				expected: overviews,
-			}),
-			Entry("should retain none with mismatching tag", testCase{
-				tags:     map[string]string{"kuma.io/service": "mobile", "version": "v2"},
-				expected: DataplaneOverviewResourceList{Items: []*DataplaneOverviewResource{}},
-			}))
-	})
-	Describe("RetainGateWayDataPlanes", func() {
-		dataplanes := DataplaneOverviewResourceList{
-			Items: []*DataplaneOverviewResource{
-				{
-					Spec: v1alpha1.DataplaneOverview{
-						Dataplane: &v1alpha1.Dataplane{
-							Networking: &v1alpha1.Dataplane_Networking{
-								Gateway: &v1alpha1.Dataplane_Networking_Gateway{
-									Tags: map[string]string{
-										"kuma.io/service": "gateway",
-									},
-								},
-							},
-						},
-					},
-				},
-				{
-					Spec: v1alpha1.DataplaneOverview{
-						Dataplane: &v1alpha1.Dataplane{
-							Networking: &v1alpha1.Dataplane_Networking{
-								Inbound: []*v1alpha1.Dataplane_Networking_Inbound{
-									{
-										Tags: map[string]string{
-											"kuma.io/service": "mobile",
-											"version":         "v1",
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		}
-		gatewayDataplanes := DataplaneOverviewResourceList{
-			Items: []*DataplaneOverviewResource{
-				{
-					Spec: v1alpha1.DataplaneOverview{
-						Dataplane: &v1alpha1.Dataplane{
-							Networking: &v1alpha1.Dataplane_Networking{
-								Gateway: &v1alpha1.Dataplane_Networking_Gateway{
-									Tags: map[string]string{
-										"kuma.io/service": "gateway",
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		}
-		It("should retain gateway overviews", func() {
-			dataplanes.RetainGatewayDataplanes("true")
-			Expect(dataplanes).To(Equal(gatewayDataplanes))
 		})
 	})
 })
